refactor: reorganize core-module
